+++ src/spycipdb/clis/cli_pre.py
@@ -41,8 +41,8 @@ from pathlib import Path
 from idpconfgen.libs.libmulticore import pool_function
 
 from spycipdb import log
-from spycipdb.components.calculators import calc_pre
-from spycipdb.components.parsers import get_exp_format_pre
+from spycipdb.core.calculators import calc_pre
+from spycipdb.core.parsers import get_exp_format_pre
 from spycipdb.libs import libcli
 from spycipdb.libs.libfuncs import get_pdb_paths
 from spycipdb.logger import S, T, init_files, report_on_crash